por hj, chega

+++ laurear/src/main/java/com/fatec/plataforma/repository/FaculdadeRepository.java
@@ -3,5 +3,5 @@ package com.fatec.plataforma.repository;
 import com.fatec.plataforma.model.FaculdadeModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FaculdadeRepository extends JpaRepository<FaculdadeModel, Long> {
+public interface FaculdadeRepository extends JpaRepository<FaculdadeModel, Integer> {
 }
